GestionDeBiblioteca/src/main/java/model/Person.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
 */
package model;

import java.util.ArrayList;
import java.util.Date;

/**
 *
 * @author monge
 */
public class Person {
    private int id;
    private Date birthDay;
    private String identification;
    private String lastName;
    private String name;
    private String phone;

    public Person() {
    }

    public Person(int id, Date birthDay, String identification, String lastName, String name, String phone) {
        this.id = id;
        this.birthDay = birthDay;
        this.identification = identification;
        this.lastName = lastName;
        this.name = name;
        this.phone = phone;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public Date getBirthDay() {
        return birthDay;
    }

    public void setBirthDay(Date birthDay) {
        this.birthDay = birthDay;
    }

    public String getIdentification() {
        return identification;
    }

    public void setIdentification(String identification) {
        this.identification = identification;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getPhone() {
        return phone;
    }

    public void setPhone(String phone) {
        this.phone = phone;
    }

    @Override
    public String toString() {
        return "Person{" + "id=" + id + ", birthDay=" + birthDay + ", identification=" + identification + ", lastName=" + lastName + ", name=" + name + ", phone=" + phone + '}';
    }

    
    
    
}
